normalisation/.ipynb_checkpoints/requirements-checkpoint.txt
fastapi
uvicorn
pandas
numpy
google-cloud-storage
scikit-learn
pyarrow
openpyxl
scipy
google-cloud-logging